[Spring 경로 조회 - 2단계] 리차드(전형중) 미션 제출합니다.

+++ src/main/java/wooteco/subway/infra/dao/SectionDao.java
@@ -34,7 +34,7 @@ public class SectionDao extends AbstractDao<SectionEntity, Long> {
                     rs.getString("up_station_name"),
                     rs.getLong("down_station_id"),
                     rs.getString("down_station_name"),
-                    rs.getInt("distance")
+                    rs.getLong("distance")
             );
 
     private final JdbcTemplate jdbcTemplate;
