+++ src/main/java/wooteco/subway/service/dto/response/PathServiceResponse.java
@@ -6,10 +6,10 @@ import wooteco.subway.domain.Station;
 public class PathServiceResponse {
 
     private final List<Station> stations;
-    private final Integer distance;
+    private final Long distance;
     private final Long fare;
 
-    public PathServiceResponse(List<Station> stations, Integer distance, Long fare) {
+    public PathServiceResponse(List<Station> stations, Long distance, Long fare) {
         this.stations = stations;
         this.distance = distance;
         this.fare = fare;
@@ -19,7 +19,7 @@ public class PathServiceResponse {
         return stations;
     }
 
-    public Integer getDistance() {
+    public Long getDistance() {
         return distance;
     }
