+++ src/main/java/wooteco/subway/ui/dto/request/SectionRequest.java
@@ -10,12 +10,12 @@ public class SectionRequest {
     @Positive
     private Long downStationId;
     @Positive
-    private int distance;
+    private Long distance;
 
     private SectionRequest() {
     }
 
-    public SectionRequest(Long upStationId, Long downStationId, int distance) {
+    public SectionRequest(Long upStationId, Long downStationId, Long distance) {
         this.upStationId = upStationId;
         this.downStationId = downStationId;
         this.distance = distance;
@@ -33,7 +33,7 @@ public class SectionRequest {
         return downStationId;
     }
 
-    public int getDistance() {
+    public Long getDistance() {
         return distance;
     }
 }
